final_hw3

+++ hw3/main_q1.cu
@@ -41,27 +41,32 @@ extern const size_t ITER_MAX_CHECK = 10;
 
 // TODO: initialize an array of size arr_size in input_array with random floats
 // between -1 and 1
-void initialize_array(vec &input_array, size_t arr_size) {
-  for (uint i = 0; i < arr_size; ++i) {
-    input_array.push_back((float)rand() / RAND_MAX * 2 - 1);
+void initialize_array(vec &input_array, size_t arr_size)
+{
+  for (size_t i = 0; i < arr_size; ++i)
+  {
+    input_array.push_back((float)rand() / (float)RAND_MAX * 2 - 1);
   }
 }
 
 void host_recurrence(vec &input_array, vec &output_array, size_t num_iter,
-                     size_t array_size) {
+                     size_t array_size)
+{
   std::transform(input_array.begin(), input_array.begin() + array_size,
-                 output_array.begin(), [&num_iter](elem_type &constant) {
+                 output_array.begin(), [&num_iter](elem_type &constant)
+                 {
                    elem_type z = 0;
                    for (size_t it = 0; it < num_iter; it++) {
                      z = z * z + constant;
                    }
-                   return z;
-                 });
+                   return z; });
 }
 
-class RecurrenceTestFixture : public ::testing::Test {
- protected:
-  RecurrenceTestFixture() {
+class RecurrenceTestFixture : public ::testing::Test
+{
+protected:
+  RecurrenceTestFixture()
+  {
     cudaFree(0);
     // initialize cuda context to avoid including cost in timings later
 
@@ -84,16 +89,17 @@ class RecurrenceTestFixture : public ::testing::Test {
 
     cudaMalloc(&device_input_array, num_bytes);
     cudaMalloc(&device_output_array, num_bytes);
-
   }
 
   // TODO: deallocate memory from both device arrays
-  ~RecurrenceTestFixture() {
+  ~RecurrenceTestFixture()
+  {
     cudaFree(device_input_array);
     cudaFree(device_output_array);
   }
 
-  void initialize() {
+  void initialize()
+  {
     initialize_array(init_arr, MAX_ARR_SIZE);
     check_initialization(init_arr, MAX_ARR_SIZE);
 
@@ -118,22 +124,26 @@ class RecurrenceTestFixture : public ::testing::Test {
 // TODO: allocate num_bytes of memory to the device arrays in the TEST FIXTURE
 // CONSTRUCTOR and deallocate the memory in the TEST FIXTURE DESTRUCTOR.
 // Hint: use cudaMalloc and cudaFree
-TEST_F(RecurrenceTestFixture, GPUAllocationTest_1) {
+TEST_F(RecurrenceTestFixture, GPUAllocationTest_1)
+{
   // if either memory allocation failed in the fixture constructor, report an
   // error message
-  if (!device_input_array || !device_output_array) {
+  if (!device_input_array || !device_output_array)
+  {
     FAIL() << "Couldn't allocate memory!" << endl;
   }
 }
 
 // TODO: Implement initialize_array function
-TEST_F(RecurrenceTestFixture, InitalizeArrayTest_2) {
+TEST_F(RecurrenceTestFixture, InitalizeArrayTest_2)
+{
   initialize_array(init_arr, MAX_ARR_SIZE);
   check_initialization(init_arr, MAX_ARR_SIZE);
 }
 
 // TODO: Implement Recurrence Kernel Fuctions in Recurrence.cuh
-TEST_F(RecurrenceTestFixture, RecurrenceKernelTest_3) {
+TEST_F(RecurrenceTestFixture, RecurrenceKernelTest_3)
+{
   // init array
   initialize();
 
@@ -150,7 +160,8 @@ TEST_F(RecurrenceTestFixture, RecurrenceKernelTest_3) {
   array_size = 1e6;
   cuda_block_size = 1024;
   cuda_grid_size = 576;
-  for (num_iter = 1; num_iter <= ITER_MAX_CHECK; ++num_iter) {
+  for (num_iter = 1; num_iter <= ITER_MAX_CHECK; ++num_iter)
+  {
     host_recurrence(init_arr, arr_host, num_iter, array_size);
     recurAndCheck(device_input_array, device_output_array, num_iter, array_size,
                   cuda_block_size, cuda_grid_size, arr_host);
@@ -160,7 +171,8 @@ TEST_F(RecurrenceTestFixture, RecurrenceKernelTest_3) {
 }
 
 // No code changes necessary
-TEST_F(RecurrenceTestFixture, RecurrenceThreadsTest_4) {
+TEST_F(RecurrenceTestFixture, RecurrenceThreadsTest_4)
+{
   /*
    * ––––––––––-------------------------------------------------------
    * Question 1.4: vary number of threads for a small number of blocks
@@ -179,7 +191,8 @@ TEST_F(RecurrenceTestFixture, RecurrenceThreadsTest_4) {
   double flops = 2 * num_iter * array_size;
   host_recurrence(init_arr, arr_host, num_iter, array_size);
   for (size_t cuda_block_size = 32; cuda_block_size <= 1024;
-       cuda_block_size += 32) {
+       cuda_block_size += 32)
+  {
     double elapsed_time =
         recurAndCheck(device_input_array, device_output_array, num_iter,
                       array_size, cuda_block_size, cuda_grid_size, arr_host);
@@ -192,7 +205,8 @@ TEST_F(RecurrenceTestFixture, RecurrenceThreadsTest_4) {
 }
 
 // No code changes necessary
-TEST_F(RecurrenceTestFixture, RecurrenceBlocksTest_5) {
+TEST_F(RecurrenceTestFixture, RecurrenceBlocksTest_5)
+{
   /*
    * ––––––––––-------------------------------------------------------
    * Question 1.5: vary number of blocks for a small number of threads
@@ -211,7 +225,8 @@ TEST_F(RecurrenceTestFixture, RecurrenceBlocksTest_5) {
   double flops = 2 * num_iter * array_size;
   host_recurrence(init_arr, arr_host, num_iter, array_size);
   for (size_t cuda_grid_size = 36; cuda_grid_size <= 1152;
-       cuda_grid_size += 36) {
+       cuda_grid_size += 36)
+  {
     double elapsed_time =
         recurAndCheck(device_input_array, device_output_array, num_iter,
                       array_size, cuda_block_size, cuda_grid_size, arr_host);
@@ -224,7 +239,8 @@ TEST_F(RecurrenceTestFixture, RecurrenceBlocksTest_5) {
 }
 
 // No code changes necessary
-TEST_F(RecurrenceTestFixture, RecurrenceItersTest_6) {
+TEST_F(RecurrenceTestFixture, RecurrenceItersTest_6)
+{
   /*
    * ––––––––––-----------------------------
    * Question 1.6: vary number of iterations
@@ -240,11 +256,12 @@ TEST_F(RecurrenceTestFixture, RecurrenceItersTest_6) {
   size_t cuda_block_size = 256;
   size_t cuda_grid_size = 576;
   size_t array_size = 1e6;
-  std::vector<size_t> num_iters = {20,   40,   60,   80,   100,  120,  140,
-                                   160,  180,  200,  300,  400,  500,  600,
-                                   700,  800,  900,  1000, 1200, 1400, 1600,
+  std::vector<size_t> num_iters = {20, 40, 60, 80, 100, 120, 140,
+                                   160, 180, 200, 300, 400, 500, 600,
+                                   700, 800, 900, 1000, 1200, 1400, 1600,
                                    1800, 2000, 2200, 2400, 2600, 2800, 3000};
-  for (size_t num_iter : num_iters) {
+  for (size_t num_iter : num_iters)
+  {
     double flops = 2 * num_iter * array_size;
     host_recurrence(init_arr, arr_host, num_iter, array_size);
     double elapsed_time =
